requirements.txt
asgiref==3.2.3
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.11
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cors-headers==3.1.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
djangorestframework==3.10.3
drf-jwt==1.13.4
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
